DTSAM-1210 Move class into testingsupport.  Add call to deprecated test for coverage.
